+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.0
+
+- Fixed a bug that caused the package to output wrong String after a chip had been deleted.
+- FEATURE ADD: Added the ability to place chips section above or below the textfield.
+- Updated library directory structure.
+
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Initial release